pyproject.toml
[project]
name = "comfyui-ugleh-openai"
version = "0.1.0"
description = "Minimal OpenAI API nodes for ComfyUI."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "ugleh" }]

dependencies = [
  "openai>=2.0.0",
  "python-dotenv",
]

[tool.setuptools]
# This is a ComfyUI custom_nodes repo; packaging is optional.
# The 'name' field is used by ComfyUI-Manager for folder normalization.

